MVS: Refactor, update CHANGELOG
