+++ src/extensions/mvs/tree/generic/tree-utils.ts
@@ -5,19 +5,19 @@
  */
 
 import { canonicalJsonString } from '../../../../mol-util/json';
-import { AdditionalProperties, DefaultsForTree, Kind, Node, SubTree, SubTreeOfKind, Tree, TreeFor, TreeSchema, TreeSchemaWithAllRequired, getParams } from './tree-schema';
+import { AdditionalProperties, DefaultsForTree, Kind, Node, Subtree, SubtreeOfKind, Tree, TreeFor, TreeSchema, TreeSchemaWithAllRequired, getParams } from './tree-schema';
 
 
 /** Run DFS (depth-first search) algorithm on a rooted tree.
  * Runs `visit` function when a node is discovered (before visiting any descendants).
  * Runs `postVisit` function when leaving a node (after all descendants have been visited). */
-export function dfs<TTree extends Tree>(root: TTree, visit?: (node: SubTree<TTree>, parent?: SubTree<TTree>) => any, postVisit?: (node: SubTree<TTree>, parent?: SubTree<TTree>) => any) {
-    return _dfs<SubTree<TTree>>(root, undefined, visit, postVisit);
+export function dfs<TTree extends Tree>(root: TTree, visit?: (node: Subtree<TTree>, parent?: Subtree<TTree>) => any, postVisit?: (node: Subtree<TTree>, parent?: Subtree<TTree>) => any) {
+    return _dfs<Subtree<TTree>>(root, undefined, visit, postVisit);
 }
-function _dfs<TTree extends Tree>(root: TTree, parent: SubTree<TTree> | undefined, visit?: (node: SubTree<TTree>, parent?: SubTree<TTree>) => any, postVisit?: (node: SubTree<TTree>, parent?: SubTree<TTree>) => any) {
+function _dfs<TTree extends Tree>(root: TTree, parent: Subtree<TTree> | undefined, visit?: (node: Subtree<TTree>, parent?: Subtree<TTree>) => any, postVisit?: (node: Subtree<TTree>, parent?: Subtree<TTree>) => any) {
     if (visit) visit(root, parent);
     for (const child of root.children ?? []) {
-        _dfs<SubTree<TTree>>(child, root, visit, postVisit);
+        _dfs<Subtree<TTree>>(child, root, visit, postVisit);
     }
     if (postVisit) postVisit(root, parent);
 }
@@ -77,15 +77,15 @@ export function copyTree<T extends Tree>(root: T): T {
  * nodes of kind `C` will be converted to `Y` with a child `Z` (original children moved to `Z`),
  * nodes of other kinds will just be copied. */
 export type ConversionRules<A extends Tree, B extends Tree> = {
-    [kind in Kind<SubTree<A>>]?: (node: SubTreeOfKind<A, kind>, parent?: SubTree<A>) => SubTree<B>[]
+    [kind in Kind<Subtree<A>>]?: (node: SubtreeOfKind<A, kind>, parent?: Subtree<A>) => Subtree<B>[]
 };
 
 /** Apply a set of conversion rules to a tree to change to a different schema. */
-export function convertTree<A extends Tree, B extends Tree>(root: A, conversions: ConversionRules<A, B>): SubTree<B> {
-    const mapping = new Map<SubTree<A>, SubTree<B>>();
-    let convertedRoot: SubTree<B>;
+export function convertTree<A extends Tree, B extends Tree>(root: A, conversions: ConversionRules<A, B>): Subtree<B> {
+    const mapping = new Map<Subtree<A>, Subtree<B>>();
+    let convertedRoot: Subtree<B>;
     dfs<A>(root, (node, parent) => {
-        const conversion = conversions[node.kind as (typeof node)['kind']] as ((n: typeof node, p?: SubTree<A>) => SubTree<B>[]) | undefined;
+        const conversion = conversions[node.kind as (typeof node)['kind']] as ((n: typeof node, p?: Subtree<A>) => Subtree<B>[]) | undefined;
         if (conversion) {
             const convertidos = conversion(node, parent);
             if (!parent && convertidos.length === 0) throw new Error('Cannot convert root to empty path');
@@ -115,13 +115,13 @@ export function convertTree<A extends Tree, B extends Tree>(root: A, conversions
 /** Create a copy of the tree where twins (siblings of the same kind with the same params) are merged into one node.
  * Applies only to the node kinds listed in `condenseNodes` (or all if undefined) except node kinds in `skipNodes`. */
 export function condenseTree<T extends Tree>(root: T, condenseNodes?: Set<Kind<Tree>>, skipNodes?: Set<Kind<Tree>>): T {
-    const map = new Map<string, SubTree<T>>();
+    const map = new Map<string, Subtree<T>>();
     const result = copyTree(root);
     dfs<T>(result, node => {
         map.clear();
-        const newChildren: SubTree<T>[] = [];
+        const newChildren: Subtree<T>[] = [];
         for (const child of node.children ?? []) {
-            let twin: SubTree<T> | undefined = undefined;
+            let twin: Subtree<T> | undefined = undefined;
             const doApply = (!condenseNodes || condenseNodes.has(child.kind)) && !skipNodes?.has(child.kind);
             if (doApply) {
                 const key = child.kind + canonicalJsonString(getParams(child));
@@ -131,7 +131,7 @@ export function condenseTree<T extends Tree>(root: T, condenseNodes?: Set<Kind<T
             if (twin) {
                 (twin.children ??= []).push(...child.children ?? []);
             } else {
-                newChildren.push(child as SubTree<T>);
+                newChildren.push(child as Subtree<T>);
             }
         }
         node.children = newChildren;
