+++ src/extensions/mvs/tree/molstar/molstar-tree.ts
@@ -1,12 +1,12 @@
 /**
- * Copyright (c) 2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2023-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Adam Midlik <midlik@gmail.com>
  */
 
 import { omitObjectKeys, pickObjectKeys } from '../../../../mol-util/object';
 import { RequiredField, bool } from '../generic/params-schema';
-import { NodeFor, TreeFor, TreeSchema } from '../generic/tree-schema';
+import { NodeFor, SubtreeOfKind, TreeFor, TreeSchema } from '../generic/tree-schema';
 import { FullMVSTreeSchema } from '../mvs/mvs-tree';
 import { MolstarParseFormatT } from '../mvs/param-types';
 
@@ -62,3 +62,6 @@ export type MolstarNode<TKind extends MolstarKind = MolstarKind> = NodeFor<typeo
 
 /** Intermediate tree representation between `MVSTree` and a real Molstar state */
 export type MolstarTree = TreeFor<typeof MolstarTreeSchema>
+
+/** Any subtree in a `MolstarTree` (e.g. its root doesn't need to be 'root') */
+export type MolstarSubtree<TKind extends MolstarKind = MolstarKind> = SubtreeOfKind<MolstarTree, TKind>
